requirements.txt
psycopg2-binary==2.8.5
requests==2.25.1
python-dotenv==0.15.0
pytz==2020.5